test: add test coverage
