+++ tests/test_llm_filter.py
@@ -1,6 +1,5 @@
 import pytest
 
-from anki_convo.card import TranslationCard
 from anki_convo.card_gen import CardGeneratorConfig
 from anki_convo.hooks.llm_filter import (
     LanguageFieldNames,
@@ -8,6 +7,7 @@ from anki_convo.hooks.llm_filter import (
     LLMFilterConfig,
     parse_llm_filter_name,
 )
+from tests.mocks import CountingCardGenerator
 
 
 def test_parse_llm_filter_name_valid():
@@ -32,30 +32,6 @@ def test_parse_llm_filter_name_invalid():
     filter_name = "invalid_filter_name"
     with pytest.raises(ValueError):
         parse_llm_filter_name(filter_name)
-
-
-class CountingCardGenerator:
-    """Card generator that returns a fixed number of cards with the same source
-    and target.
-
-    It keeps track of how often it was called in order to test that lru_cache works.
-    """
-
-    def __init__(self, n_cards: int = 1):
-        self.n_cards = n_cards
-        self.n_times_called = 0
-
-    def __call__(self, card: TranslationCard):
-        self.n_times_called += 1
-        source = f"Source of card for {card} after {self.n_times_called} call(s)"
-        target = f"Target of card for {card} after {self.n_times_called} call(s)"
-
-        return [
-            TranslationCard(
-                source=f"{source} (card {i})", target=f"{target} (card {i})"
-            )
-            for i in range(self.n_cards)
-        ]
 
 
 @pytest.fixture()
